ENH: Read "src" property in config
This commit adds support for a lot of things because all the passed options
to another modules should be allowable to be modified from the config file.
For example, we need cwd for child_process.exec, but also for chokidar files.
So, setting options in two commits to separate those things gets more
complicated.. Anyway, a config could be passed like so: {
  "precise-watcher": {
    // expects an object to pass global chokidar options.
    "chokidar": {},
    // expects an array, but should handle any value.
    "src": [{
      // expects an array, but should handle any value.
      "pattern": [/* glob pattern */],
      // expects an string. Replaces src.regexp matches.
      "run": "echo <file> && echo <file>", /** example */
      // expects an object to pass local chokidar options.
      "chokidar": {},
      // expects any chokidar event, like "add", "change", ...
      // defaults to "change".
      "on": "change",
      // expects and string or a regexp to replace tokens in your command.
      "regexp": /<file>/g
    }]
  }
}

TST: Add integration test for "src" property in integration/src.test.js

TST: Set a valid config in 1st test in integration/src.test.js

A boolean is not a valid config. Set an object instead.

ENH: .gitignore temp*